fix: move writer to io
